FLAC support added
FLAC was a pain in the neck. First because the library isn't nearly as
easy to use as Ogg Vorbis, and secondly because I made a maddening bug
for myself.

And supporting FLAC wasn't even the original plan, I only did it
because Ogg Vorbis was so easy and only took a couple hours. No good
deed goes unpunished.